+++ apps/backend/api/graphql/mutations/index.js
@@ -18,6 +18,12 @@ export {
   removePostFromCollection
 } from './collection'
 export {
+  grantContentAccess,
+  revokeContentAccess,
+  refundContentAccess,
+  recordStripePurchase
+} from './contentAccess'
+export {
   createComment,
   createMessage,
   deleteComment,
@@ -164,6 +170,17 @@ export {
   deleteZapierTrigger
 } from './zapier'
 export {
+  createStripeConnectedAccount,
+  createStripeAccountLink,
+  createStripeOffering,
+  updateStripeOffering,
+  createStripeCheckoutSession,
+  checkStripeStatus
+} from './stripe'
+export {
+  membershipChangeCommit
+} from './membershipChange'
+export {
   addEmailEnabledTester,
   removeEmailEnabledTester
 } from './emailEnabledTesters'
@@ -268,6 +285,12 @@ export function markAllActivitiesRead (userId) {
 export async function markThreadRead (root, { messageThreadId }, context) {
   const messageThread = await Post.find(messageThreadId)
   await messageThread.markAsRead(context.currentUserId)
+  return messageThread
+}
+
+export async function markThreadUnread (root, { messageThreadId }, context) {
+  const messageThread = await Post.find(messageThreadId)
+  await messageThread.markAsUnread(context.currentUserId)
   return messageThread
 }