+++ apps/backend/api/models/EventInvitation.js
@@ -23,7 +23,8 @@ module.exports = bookshelf.Model.extend({
   },
 
   incrementIcalSequence: async function () {
-    return this.save({ ical_sequence: this.getIcalSequence() + 1 })
+    const nextSeq = this.getIcalSequence() + 1
+    return this.save({ ical_sequence: nextSeq }, { patch: true })
   },
 
   notGoing: function () {