+++ apps/web/src/store/actions/fetchThreads.js
@@ -2,20 +2,21 @@ import { get } from 'lodash/fp'
 import MessageThreadsQuery from '@graphql/queries/MessageThreadsQuery'
 import { FETCH_THREADS } from 'store/constants'
 
-export default function (first = 10, offset = 0) {
+export default function (first = 10, offset = 0, search) {
+  const variables = { first, offset }
+  if (search) variables.search = search
+
   return {
     type: FETCH_THREADS,
     graphql: {
       query: MessageThreadsQuery,
-      variables: {
-        first,
-        offset
-      }
+      variables
     },
     meta: {
       extractModel: 'Me',
       extractQueryResults: {
-        getItems: get('payload.data.me.messageThreads')
+        getItems: get('payload.data.me.messageThreads'),
+        getRouteParams: () => (search ? { search } : {})
       }
     }
   }